переход на MVVM и coroutines

+++ app/src/main/java/com/example/weatherforecasts/WeatherList/RecViList/RecViHourItem/Item.kt
@@ -2,7 +2,7 @@ package com.example.weatherforecasts.WeatherList.RecViList.RecViHourItem
 
 data class Item(
     val temperatureAir: Int,
-    val prec_mm: Int,
+    val prec_mm: Float,
     val icon: String? = null,
     val date: String? = null,
 )
